requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
pydantic
requests
